move receivemodal to own folder
