+++ src/pages/Home/ReceiveModal/__tests__/ReceiveModal.test.tsx
@@ -1,8 +1,8 @@
 import { render, screen } from "@testing-library/react";
 import userEvent from "@testing-library/user-event";
 import { I18nextProvider } from "react-i18next";
-import i18n from "../../../i18n/test_config";
-import { rest, server } from "../../../testServer";
+import i18n from "../../../../i18n/test_config";
+import { rest, server } from "../../../../testServer";
 import ReceiveModal from "../ReceiveModal";
 
 beforeAll(() => {
