+++ src/pages/Home/TransactionCard/__tests__/TransactionCard.test.tsx
@@ -1,8 +1,8 @@
 import { render, screen } from "@testing-library/react";
 import { I18nextProvider } from "react-i18next";
-import i18n from "../../../i18n/test_config";
-import type { Props } from "./TransactionCard";
-import TransactionCard from "./TransactionCard";
+import i18n from "../../../../i18n/test_config";
+import type { Props } from "../TransactionCard";
+import TransactionCard from "../TransactionCard";
 
 const props: Props = {
   transactions: [],
